# docs/API/openapi-schema.yaml
openapi: 3.0.3
info:
  title: ApexStore Mini E-Commerce & Order Management API
  version: 1.0.0
  description: Comprehensive REST API documentation for ApexStore e-commerce platform.
    Covers Authentication, Products, Categories, Cart, Wishlist, Coupons, Orders,
    Payments, Inventory, and Admin Analytics.
paths:
...
  /api/coupons/{id}/:
    get:
      operationId: coupons_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - coupons
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
          description: ''
    put:
      operationId: coupons_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - coupons
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Coupon'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Coupon'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Coupon'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
          description: ''
    patch:
      operationId: coupons_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - coupons
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCoupon'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCoupon'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCoupon'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coupon'
          description: ''
    delete:
      operationId: coupons_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - coupons
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        code:
          type: string
          maxLength: 50
        description:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountTypeEnum'
        discount_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        min_order_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        max_discount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        start_date:
          type: string
          format: date-time
        expiration_date:
          type: string
          format: date-time
        usage_limit:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
          nullable: true
          description: Total system usage limit
        per_user_usage_limit:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        is_active:
          type: boolean
        total_used:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - code
      - created_at
      - discount_value
      - expiration_date
      - id
      - total_used
      - updated_at
...
    DiscountTypeEnum:
      enum:
      - PERCENTAGE
      - FIXED_AMOUNT
      type: string
      description: |-
        * `PERCENTAGE` - Percentage
        * `FIXED_AMOUNT` - Fixed Amount
...
    PatchedCoupon:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        code:
          type: string
          maxLength: 50
        description:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountTypeEnum'
        discount_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        min_order_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        max_discount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        start_date:
          type: string
          format: date-time
        expiration_date:
          type: string
          format: date-time
        usage_limit:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
          nullable: true
          description: Total system usage limit
        per_user_usage_limit:
          type: integer
          maximum: 4294967295
          minimum: 0
          format: int64
        is_active:
          type: boolean
        total_used:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT